move EmitterOptions to options.ts
