+++ packages/http-client-java/emitter/src/index.ts
@@ -1,4 +1,5 @@
 // $onEmit is the entry of the emitter
 export { $onEmit } from "./emitter.js";
 // emit options interface is required to be exported
-export { $lib, DevOptions, EmitterOptions } from "./lib.js";
+export { $lib } from "./lib.js";
+export { DevOptions, EmitterOptions } from "./options.js";
